+++ bin/blueprint.php
@@ -1,11 +1,11 @@
 <?php
 
 // Includes the Composer Autoloader
-require realpath('vendor') . '/autoload.php';
+require 'vendor/autoload.php';
 
 if (! defined('BLUEPRINT_FILENAME')) {
     define('BLUEPRINT_FILENAME', 'blueprint.yml');
 }
 
 // Prepares the console and runs the application
-Rougin\Blueprint\Console::boot(BLUEPRINT_FILENAME)->run();
+Rougin\Blueprint\Console::boot(BLUEPRINT_FILENAME, new Auryn\Injector)->run();